+++ include/utilities.hpp
@@ -3,12 +3,13 @@
 
 #include <cmath>	// for mathematical functions
 #include <vector>	// for std::vector
+#include "vector.hpp"
 
 // a place to aggregate helper functions which do not fit anywhere else
 namespace utilities {
 
-	void fillValuesPowerLaw(const double minValue, const double maxValue, std::vector<double>& values, const int power);
-	void fillValuesLogarithmic(const double minValue, const double maxValue, std::vector<double>& values);
+	void fillValuesPowerLaw(const NUMERIC minValue, const NUMERIC maxValue, std::vector<NUMERIC>& values, const int power);
+	void fillValuesLogarithmic(const NUMERIC minValue, const NUMERIC maxValue, std::vector<NUMERIC>& values);
 }
 
 #endif
