+++ src/vector.cpp
@@ -1,7 +1,7 @@
 #include "vector.hpp"
 
 // constructor
-vector::vector(std::initializer_list<double> list) : vector(list.size())  {
+vector::vector(std::initializer_list<NUMERIC> list) : vector(list.size())  {
         for(unsigned i = 0; i < list.size(); ++ i)
             (*this)[i] = *(list.begin() + i);
 }
